requirements.txt
streamlit
feedparser
python-dotenv
google-genai